+++ src/main/java/com/catalyst/collector/services/impl/CollectionsServiceImpl.java
@@ -38,10 +38,6 @@ public class CollectionsServiceImpl implements CollectionsService {
 	}
 
 
-	@Override
-	public void addAge(Age age) {
-		collectionsDao.addAge(age);
-	}
 
 	@Override
 	public ArrayList<Age> getAgeTypes(){
@@ -49,14 +45,37 @@ public class CollectionsServiceImpl implements CollectionsService {
 	}
 
 	@Override
-	public void updateAge(Age age){
-		collectionsDao.updateAge(age);
+	public boolean addAge(Age age) {
+		String ageString = age.getAge();
+		if(ageString.length() <= 255 && !ageString.matches(".*\\d.*")) { //Maximum of 255 characters for an age, no digits allowed
+			collectionsDao.addAge(age);
+			return true;
+		}
+		return false;
+	}
+
+
+	@Override
+	public boolean updateAge(Age age){
+		String ageString = age.getAge();
+		if(ageString.length() <= 255 && !ageString.matches(".*\\d.*")) { //Maximum of 255 characters for an age, no digits allowed
+			collectionsDao.updateAge(age);
+			return true;
+		}
+		return false;
+
 	}
 
 	@Override
-	public void deleteAge(Integer id) {
-		collectionsDao.deleteAge(id);
+	public boolean deleteAge(Integer id) {
+		if(id > 0) {
+			collectionsDao.deleteAge(id);
+			return true;
+		}
+		return false;
 	}
+
+
 	public ArrayList<Category> getCategory() {
 
 		return collectionsDao.getCategory();
